firmware_Cr50RMAOpen: use ccd_lockout
We can't open cr50 if ccd is locked out. Modify init to check
ccd_lockout before trying to open cr50. If we can't set all capabilities
to default during init, raise an error before running the test.

BUG=none
BRANCH=none
TEST=run firmware_Cr50RMAOpen on mp and premp image. Set a capability to
Always and make sure the test fails if it can't set it to Default.

Change-Id: I48ac6a9e61c5db43100c6af4ff77d8d01216f23b
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1096364
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>