faft: Make USB enable signal naming more flexible
Not all boards use USB%d_ENABLE as the enable signal for the USB-A
ports. This change allows other names to be used so that the EC source
can match the net name used in schematics and layouts.

This commit additionally updates the FAFT values for Octopus to
explicitly use this change.

BUG=b:79872754
TEST=test_that <ip> firmware_ECUsbPorts on yorp proto1
Change-Id: I0dfad8f80a7b690698ffe0a1797530152d89c170
Reviewed-on: https://chromium-review.googlesource.com/1096124
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>