Add more cr50 suites
Create a pvt and prepvt test suite for cr50. The images are slightly
different so we need to run the prepvt and pvt suites with different
args.

We will remove cr50_stress_experimental once the prepvt and pvt suites
are verified.

BUG=none
BRANCH=none
TEST=run prepvt and pvt images. verify ccd_lockout True for pvt and
False for prepvt

Change-Id: Iff7df1361503e73d6148e345c4783d6438eea721
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1100046
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>