build_externals: announce success moofully.
build_externals (by design) emits a bunch of errors to stdout
on a successful run.  This contributes to the widespread
problem of ChromeOS infrastructure not making success or failure
obvious.  In order to make success/failure more clear here,
this change fixes this problem by announcing success for all
to see.

cowsay, one of the greatest unix tools of all time, can be
used for good or evil.  In this case, it is used for good.

BUG=None
TEST=Run build_externals.py

Change-Id: I5bde86bb847d3a455c234815decf1f5a9faa5235
Reviewed-on: https://chromium-review.googlesource.com/1099732
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Craig Bergstrom <craigb@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>