platform_Flashrom: account for unibuild firmware tarballs
The platform_Flashrom test at one point tries to reprogram the DUT
flash device with the BIOS image from chromeos-firmwareupdate self
extracting archive.

The test expects to see the bios.bin file in the top level directory
created when the tarball is unpacked.

With recent introduction of the unibuild there could be multiple bios
images in the same tarball, located in the models subdirectory.

This patch modifies the test to use the FirmwareUpdater() class to
properly locate the bios image matching the model the test is running
on.

Also removed the unused re import, as suggested by the linter.

BRANCH=none
BUG=b:79492818
TEST=verified that platform_Flashrom test passes on Fizz(unibuild) and
     Pyro(single bios.bin).

Change-Id: I1787c8c14c69f8da4860224b93d62b52ed6e4f7d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1100492
Reviewed-by: Wai-Hong Tam <waihong@google.com>