[autotest] Use 'repair' in place of 'verify' in deployment.
At the end of a `deploy` command, we need to schedule a special task
on the successful DUTs in order to get their status up-to-date.
This changes the special task from repair to verify to deal with
certain edge cases relating to getting labels applied correctly.

BUG=chromium:850729
TEST=None

Change-Id: I60dc61f4f82768c53fcdbd735a79e418284907f7
Reviewed-on: https://chromium-review.googlesource.com/1102356
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>