[autotest] Use fping for moblab DUT discovery.
arp or ip n commands seem to be unreliable at times, showing
phantom IP's when using cisco managed router in the lab but showing
IP addresses as DELAY rather than REACHABLE in the VM.

Switch to a new strategy, just ping the devices and include devices
that have a 0% ping loss.

The change drops detection of the mac address, there is no known
need for it.  Also moblab should always have fping installed it is
in the moblab ebuild as an RDEPEND.

Tryjobs passed guado_moblab-paladin,
https://ci.chromium.org/p/chromeos/builders/luci.chromeos.general/Try/b8943249388377048672

moblab-generic-vm-paladin
https://ci.chromium.org/p/chromeos/builders/luci.chromeos.general/Try/b8943249380296023168

BUG=chromium:853337
TEST=trybot

Change-Id: I96bdf3c70b42081225bbbff69b519216ce166fd9
Reviewed-on: https://chromium-review.googlesource.com/1107238
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>