[autotest] Handle version labels from `cros stage`
The tool `cros stage` updates and uploads test image content to a
new location in googlestorage.  In uploading to a new location, the
process invalidates the value of CHROMEOS_RELEASE_BUILDER_PATH in
/etc/lsb-release in the root file system.

Provisioning believes that the URL from which the image is
downloaded and the setting of CHROMEOS_RELEASE_BUILDER_PATH are
matched, and includes a sanity check to make sure the image
installed has the expected setting.  Images from `cros stage`
break this check, resulting in failed provisioning tasks.

This amends the check to exclude the specific label names
produced by `cros stage` from the provisioning check.

BUG=chromium:844276
TEST=None

Change-Id: Ic7865f804fda2fcbed4711441806dbabeff486df
Reviewed-on: https://chromium-review.googlesource.com/1115715
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Laurence Goodby <lgoodby@chromium.org>