firmware_LockedME: accomodate new flashrom behavior
Recent changes to the CrOS version of flashrom result in writes to
flash regions protected by the PCH (when instructed by the contents of
SI_DESC) to not result in a failed exit status, but still fail to
write* (good) and log an access failure (observed on both an Intel big
core SoC and Intel small core SoC).

The CrOS flashrom maintainer should make some level of effort to couple
flashrom behavior changes with faft: so, if the behavior changes back to
an exit status of zero, this change should be reverted at the same time
(it should be innocuous in that case, but to avoid build up of detritus
in this test).

*fail to write observation is based on the log of the access failure and
read attempts of protected regions still being all ones.

BUG=b:110839551, b:110572221
TEST=Ran against a small core DUT (epaullete) with a locked TXE (passed
as would be expected).  Ran against a big core DUT (pantheon) with the
manufacturing mode check commented out (failed as expected with SI_DESC
detected as writeable).

Change-Id: I83ec216fcb6018961077a3052d54b771a2ec4182
Reviewed-on: https://chromium-review.googlesource.com/1116192
Commit-Ready: Kevin Shelton <kmshelton@chromium.org>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>