[autotest][cfm] Converts the media stats timestamp to seconds on the JS side
crbug.com/851482 introduced an issue that caused the media stats timestamps to
always be set to -2147483648. By converting the timestamp to seconds on the
JS side (instead of processing devtools response) we can work around this
regression.

Note that this only fixes timestamps. Other stats that exceed the int range
will still be set to -2147483648 when queried over devtools. From my analysis
we do not store any of these in the perf dashboard - if we do I will create
follow up CLs for them.

BUG=chromium:851482, chromium:850880
TEST=Tested with local setup, timestamps return same values again.

Change-Id: I44ff9ac42e4dd5e6ff26919ec033bc8b937b92a5
Reviewed-on: https://chromium-review.googlesource.com/1116547
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Claes Malmnäs <malmnas@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>