faft: Check is_firmware_saved before calling restore_firmware
is_firmware_changed called by restore_firmware will throw an exception
if the firmware has not been saved. If the test happens to fail before
the firmware has been backed up we should not throw an exception in the
cleanup method. This just adds noise to the already verbose autotest
logs.

BUG=b:110817463
BRANCH=none
TEST=Ran firmware_CorruptFwSigB and verified that I no longer get an
exception restoring the firmware when the test failed before the backup.

Change-Id: Idbeaf457b4af9b146e7e47b2bb627bf501355a63
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1118923
Reviewed-by: Jett Rink <jettrink@chromium.org>