[autotest] Set a "target version" for stateful updates
In some cases in the lab, provisioning may downgrade a DUT.  Recent
OS changes have introduced a problem where downgrades from builds
after R69-10756.0.0 to builds before may fail and cause a powerwash.

This changes the provisioning flow to create a file in stateful that
indicates the target version of an update.  Shutdown code in the OS
uses this file to recognize when a downgrade is occurring, and
prevent the unwanted powerwash.

BUG=chromium:854278
TEST=Run sanity suite on a local Autotest instance; examine logs

Change-Id: I757118ac94a6f4e590a961b654cd24fa220d633b
Reviewed-on: https://chromium-review.googlesource.com/1119197
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>