control.crosbolt_perf_weekly: Fix timeout value
The time elapsed in autotest shard starts from the job scheduled,
not from the running of an autotest. Therefore, we have to relax
the timeout_mins as a weekly job could be scheduled in the middle
of perbuild jobs and could not be finished in four hours.

BUG=chromium:859767
TEST=none

Change-Id: Ia79a64e1925c2574afe1bbf49d3a7d517320b2a2
Signed-off-by: Chung-yih Wang <cywang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1123965
Commit-Ready: Chung-yih Wang <cywang@chromium.org>
Tested-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Hsu-Cheng Tsai <hctsai@chromium.org>