autotest: Add retry for database connections
Django manages connections internally for its ORM.

Since this can occur on basically any ORM interaction, it's
unreasonable to try to wrap every such call site with try/except.
Django itself does not provide a way to retry on these
errors (probably because for a request based model, intermittent
database connection failures are fairly harmless).

So the best approach is to add a retry for our custom database
backend.

This hack will need to be kept so long as we keep using the Django ORM
beyond request scope, or if the failure of a whole process
occasionally can be tolerated.  The ideal solution would be to have an
RPC backend abstracting away the database for job processes.  The job
processes would then see an occasional RPC error when a connection
error kills a single RPC request.

BUG=chromium:805724
TEST=None

Change-Id: Ia4de71b7c23f9dc37bf9d6389dc6b394638eef22
Reviewed-on: https://chromium-review.googlesource.com/1128389
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>