autotest: Check tags instead of bot_id when resuming provision suite
It's possible that when we resume a suite, its child task is still
pending. In this case, the json info of a retrieved task is like:
{
    "created_ts": "2018-07-12T21:07:01.156070",
    "current_task_slice": "1",
    "name": "dummy_Pass",
    "server_versions": [
      "3675-21ffa58"
    ],
    "state": "PENDING",
    "tags": [
      "build:nyan_blaze-paladin-tryjob/R69-10869.0.0-b2743183",
      "id:chromeos-skylab-bot-140e9f86-ffef-49ea-bb07-40494e0b0481",
      ...
    ],
    "task_id": "3ea9231703a70010",
    "user": "xixuan@google.com"
}

It doesn't include 'bot_id', only include 'name' and 'tags'. This CL uses tags
to match the bot_id of each child task of a provision suite.

BUG=chromium:860027
TEST=Ran "./bin/run_suite_skylab --build
nyan_blaze-paladin-tryjob/R69-10869.0.0-b2743183 --board nyan_blaze
--suite_name provision --pool suites --priority 50 --timeout_mins 180
--test_retry --max_retries 5 --suite_args "{u'num_required': 1}"
--suite_id 3ea922fe44959f11" locally.

Change-Id: Ibe49be87b47f621355256324b53cc032dd165197
Reviewed-on: https://chromium-review.googlesource.com/1135790
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>