firmware_Cr50Password: add more test cases
It matters which state you're in when you set the password. If you set
the password in open, you can only clear the password in open. If you
set the password in unlock, you can clear the password in open or
unlock.

BUG=none
BRANCH=none
TEST=run with 0.4.8

Change-Id: I0d865468f18f1d1b83ee2bc24b6248f3c576653c
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1135919
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>