network_WiFi_UpdateRouter: Modify update router test to include packet capture
In setups that use whirlwind routers (> 60%), we also have a second
whirlwind that is used as packet capture which also needs to be
updated to the latest whirlwind images when available. This CL
modified the network wifi update router test to acheive this.

BUG=chromium:863595
TEST=Tested against 2 different hosts (with and without separate pcap)
in lab.

Change-Id: Ib4c5b2cd98b5c09549a3994b7f822cffadb3e9e6
Reviewed-on: https://chromium-review.googlesource.com/1137066
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>