[Autotest] Add test for DeviceAutoUpdateDisabled policy
Add an AU test which checks the behavior of the
DeviceAutoUpdateDisabled device policy.

Server test wrapper policy_AUServer:
- clears the TPM (both before and after the test to ensure
  no lab machines are left in an enrolled state)
- stages a payload in a reachable location
- kicks off the client test

Client test policy_DeviceAutoUpdateDisabled:
- fake-enrolls the device and sets the policy as desired
- sets up a NanoOmaha server to respond to update requests
- verifies that update request is or is not sent, as per policy
- verifies that update does or does not start, as per policy

There are three control files, each with a different policy value.
When the policy is set to True, update is disabled and no update
requests are sent.  When the policy is set to False or not set at
all (here None is used to represent that), update is not disabled
and device can update.

enterprise_au_context adds some helpful functions that are useful
for policy AU tests but are likely too specific for
update_engine_util.

The enterprise_policy_base changes are to handle the difference
between what the FakeDMS expects and the policy name from the
policy template (which is also what is listed in chrome://policy).
This dict will not be difficult to maintain in the long term.

TEST=ran on several lab machines with no issue
BUG=None

Change-Id: Iee6cf840e09e0705b166636dbf896c46d3382662
Reviewed-on: https://chromium-review.googlesource.com/1144536
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>