autotest: Use /tmp for SSH sockets
Previously, we use the "default temp directory".  This is platform
dependent, but crucially can vary based on TMPDIR, TEMP, or TMP
environment variables.  I'm pretty sure Swarming sets this to a
different directory for each task.

The problem is that paths for Unix sockets are limited to 108 bytes,
not the standard 4096 for regular paths.

The Python documentations says:

The default directory is chosen from a platform-dependent list, but
the user of the application can control the directory location by
setting the TMPDIR, TEMP or TMP environment variables. There is thus
no guarantee that the generated filename will have any nice
properties, such as not requiring quoting when passed to external
commands via os.popen().

The key is that there is no guarantee that the path can fit
comfortably in 108 bytes either.

So we should hard code to use /tmp instead of relying on the default.

Note that this does make us dependent on POSIX platforms, but Autotest
isn't really portable anyway.

BUG=chromium:865171
TEST=None

Change-Id: I79af4285e055846a87b37e35bdb0ce6957c07417
Reviewed-on: https://chromium-review.googlesource.com/1144404
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>