autotest: Log the link of the swarming task to wait for a suite
If run_suite_skylab is kicked off with a passed-in suite_id, the process
is designed to resume the suite and wait for it to finish. In this case,
it's not enough to only log the link of the task in which the suite is
created. It's also needed to log the link of the task in which the suite
is waited for.

E.g.
https://uberchromegw.corp.google.com/i/chromeos/builders/nyan_blaze-paladin/builds/413

It only records
https://chrome-swarming.appspot.com/task?id=3ece559a9ae6b411&refresh=10&show_raw=1.

It should also records
https://chrome-swarming.appspot.com/task?id=3ece55df8a4dd310&refresh=10

BUG=chromium:854426
TEST=Ran it on cros-skylab-suite-server1.hot.

Change-Id: I49360c6a4860453723849e183809f6d8c13cc358
Reviewed-on: https://chromium-review.googlesource.com/1145944
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>