autotest: wifi: perf test with specific governor
Our network_WiFi_Perf tests run with whatever cpu governor is set on the
DUT and router. This change allows for perf tests to be run in which one
set of tests is performed with the existing cpu governors, and the other
set of tests to be run with the specified governer applied to all cores
on both the DUT and router.

By default, the tests remain unchanged. One must run the test with
'--args="governor=$(GOVERNOR_NAME)" to get the new functionality.

BUG=chromium:860844
TEST=Ran `test_that -b $(BOARD) $(MACHINE) network_WiFi_Perf.vht80 to
     ensure test will, by default, work as before (except for the added
     'governor-default' part of the test tag).
TEST=Ran `test_that -b $(BOARD) $(MACHINE) network_WiFi_Perf.vht80
     --args="governor=performance"` to take the performance test with
     the default and performance governors.

Change-Id: Ie735cea67f8528e2f715887b560a9679a85e0044
Reviewed-on: https://chromium-review.googlesource.com/1148584
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>