autotest: add graphics_HwOverlay tests
This CL adds a new autotest for checking if 2D/3D HTML canvases
are correctly promoted to overlays. Counting the number of
overlays is added to the graphics_utils.py file; the rest is
a combination of a py file and control and html files.

Note that it needs a flag to work on ToT, which is passed to
Chrome, and that it should work on Rockchip (w/o screen rotation)
but not on Intel (WIP).

crrev.com/c/1144481 plugs it into autotest-chrome.

Bug: 865112
Change-Id: I22d71cf45799378e2574c33b0d389348711b0d23
Reviewed-on: https://chromium-review.googlesource.com/1142349
Commit-Ready: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>