add test to verify erasing of various block sizes
Recent flashrom changes ensure that the most appropriate flash block
erase commands are used depending on the size of the area which needs
to be erased.

This caused a lot of issues with controllers where not all block erase
sizes supported by the chip are supported by the controller.

The new test finds the inactive AP RW firmware section and tries
erasing there blocks of various sizes.

BRANCH=none
BUG=chromium:860142
TEST=tried running on bob, cyan, and eve:

    test_that --board=<BOARD> vb-03.mtv platform_FlashErasers

  succeeded on all devices.

Change-Id: I925b8efb9219ffbbc547c259330e91bc4b356fd2
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1149502
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>