firmware_DevMode: Check if EC SW Sync is disabled.
firmware_DevMode tests that entering dev mode while the EC is in RW does
not work, however in order for the EC to be in RW, EC software sync
should be disabled.  This commit adds a check to see whether or not the
"disable EC software sync" GBB flag is set and if it is, skips the test
case of entering dev mode while EC_IN_RW=1.

BUG=b:111878226
TEST=Run firmware_DevMode against Nocturne with EC software sync
disabled (via "no_ec_sync" arg) and enabled.  Verify that the test
passes in both cases.

Change-Id: If9db3abcd214308bac46a9f937ddbf34b91b1142
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1152185
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>