firmware_IntegratedU2F: wait until owner key exists
Wait a bit for the owner key to be created if it doesn't exist. It
shouldn't take very long, but if it isn't created, this will make the
failure more obvious. Raise an error if cryptohome isn't running or if
owner.key wasn't created.

BUG=b:78589631
BRANCH=none
TEST=run firmware_IntegratedU2F a bunch

Change-Id: I836921a6d91a21731a053682fd2a980260c3c34d
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1152176
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>