[autotest] network_WiFi_MaskedBSSID: stop requiring MULTI_AP_SAME_BAND
This test has some strange requirements: it wants to set up an illegal
configuration, with 2 BSS's using the same BSSID, to imitate some broken
routers in the field. The Linux mac80211 framework doesn't accept this,
returning -ENOTUNIQ instead, so this doesn't work when you run both of
these BSS's on the same radio.

This all worked OK on APs that had more than 1 radio for each band (so
you work around Linux's per-interface BSSID restriction), but it doesn't
work on Gale, where we force the 2 BSS's onto the same radio.

We can work around this by just switching this test to put the two
incompatible BSS's on separate bands (2.4GHz / 5GHz), and then drop the
CAPABILITY_MULTI_AP_SAME_BAND requirement.

This should also fix some issues seen on some Whirlwind routers, where
the 3rd radio (phy2) wasn't operating reliably. We may try to avoid
using this radio entirely in the future, so this is a good start.

BUG=chromium:774808, chromium:866181
TEST=network_WiFi_MaskedBSSID.wifi_masked_bssid on Gale

Change-Id: Idab9c56f42ad426a0f6b323e49539699679cd2d4
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1159461
Reviewed-by: Grant Grundler <grundler@chromium.org>