[autotest] network_WiFi_VerifyRouter: don't set antenna configuration twice
We set the antenna mask to 0x3 to start off, since this test used to be
assuming we had 2 antennas on all radios. But then, it re-set the
antenna configuration later to the requested bitmap.

Let's just do this once.

Also, we might as well stick the magic delay after all the antenna
configuration is done.

BUG=none
TEST=network_WiFi_VerifyRouter

Change-Id: I663949c1ef5d4372965f7f7d2dc08c64ea485eb4
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1162967
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>