[autotest] network_WiFi_BgscanBackoff: straighten out router requirements
As written today, this test requires CAPABILITY_MULTI_AP_SAME_BAND but
doesn't declare it. The test starts up multiple BSS's at distinct
frequencies, which can't be served by a single radio.

For the 'wifi_bgscan_backoff' test, this isn't really required; the test
can just as well be run on separate 2G vs. 5G bands.

For the '5760noise_check' variant, we explicitly wanted to test two
5GHz channels. This isn't possible on Gale, so let's add a capability
check so this test gets a TEST_NA result.

As a related effect, this also should move the .wifi_bgscan_backoff
variant to avoid running on phy2 on Whirlwind, which can help avoid some
flakiness. Whirlwind's phy2 is not known to be a reliable transmitter,
and we may stop using it entirely soon.

BUG=chromium:774808, chromium:866181
TEST=network_WiFi_BgscanBackoff.wifi_bgscan_backoff
     and network_WiFi_BgscanBackoff.5760_noise_check with gale;
     the former now passes, and the latter gets TEST_NA

Change-Id: I9f6d7ea0dba86d84aaa8cbc8ca236baf8fbdf92b
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1159462
Reviewed-by: Grant Grundler <grundler@chromium.org>