Reland "security_SandboxedServices: reflect shill sandboxing changes"
This is a reland of a248a9ab77c19be3afff948de8151e03b25fd619.

The only reason we reverted the change was to merge the revert back into
M69, but we still want the changes in TOT/M70 -- hence the reland.

CQ-DEPEND=CL:1162311
BUG=chromium:649417
TEST=This test was passing in TOT/M70 yesterday before revert, so just
going back to that state.

Change-Id: I3802f4ae923ddf08d44926863cbdd813568aa3ef
Reviewed-on: https://chromium-review.googlesource.com/1162312
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>