[autotest] network_WiFi_BgscanBackoff: stop using Whirlwind's phy2
The 5760noise_check variant of this test was requesting 1 spatial
stream, so that it could run on Whirlwind's phy2. Don't do this, because
phy2 is not normally used in production as a transmitter, and because
our lab conductive setups don't usually wire up its antennas.

As an effect of this, we can't support 2 simultaneous channels on the 5
GHz band. Just use different channels from the 2.4 and 5 GHz bands.

Per Kirtika's suggestion, I make the 5760noise_check variant roughly
comparable to the wifi_bgscan_backoff variant, so that we can see
whether channel 153 (a known noisy channel) behaves significantly
differently than a known less-noisy 5GHz channel.

BUG=chromium:774808, chromium:866181
TEST=network_WiFi_BgscanBackoff.wifi_bgscan_backoff
     and network_WiFi_BgscanBackoff.5760_noise_check with gale;
     both now pass; run w/ whirlwind, and see we avoid phy2

Change-Id: I22227cee072d362ceb00bda39876717a374a1d44
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1162983
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>