autotest: Add 'use-fallback' to run_suite_skylab
For release builder or suite-scheduler, normal suites will be scheduled
on DUTs, but there's no provision stage executed on them before.
So we want to first schedule on DUTs with the given build. If there's no
such DUTs, fallback to schedule it on other DUTs.

BUG=chromium:871440
TEST=Run it on cros-skylab-suite-server1.hot.

Change-Id: I029392cf9cb3ff69fd0bb71302ad522f3f859972
Reviewed-on: https://chromium-review.googlesource.com/1164541
Commit-Queue: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>