autotest: Improve tast.py error logging.
Make the tast.py server test log the tast process's first
line of stderr on failure if stdout is empty. The "list"
subcommand logs to stderr instead of stdout (since it writes
tests to stdout), so this produces a better failure message
if the DUT is initially unreachable.

BUG=chromium:778389
TEST=updated tests

Change-Id: I4f7244191303d2405c39ce76902b820d7306739b
Reviewed-on: https://chromium-review.googlesource.com/1172058
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>