autotest: Stage control file inside SSP container for skylab
In skylab, control file name is passed in and autoserv finds the control
file to execute using the name.

When running a test with SSP, autoserv re-executes inside the container
with somewhat modified commandline. For skylab, autoserv now passes in
--control-file unchanged into the container. autoserv execution inside
the container finds and stages the control file. Because the autotest
checkout inside the container is from the test artifacts for the target
build, this results in the control file being staged from the right
build.

BUG=chromium:871439
TEST=Manual.

Change-Id: Ide2b1173a40b08bfc598537c15731d613da3eb19
Reviewed-on: https://chromium-review.googlesource.com/1172070
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>