autotest: Drop support for drones without SSP
This CL removes the --warn-no-ssp argument passed to autoserv and all
the logic around creating that argument.
This argument would make autoserv spew a warning in the logs and ignore
the --ssp-required option. But,
[1] All drones have SSP enabled (have had for years?)
[2] This behaviour is dangerous. Nobody checks the logged warnings and
    users will assume their test ran with SSP.

This argument probably made sense when we were slowly rolling out SSP,
but we're better off without it now.

BUG=None
TEST=None

Change-Id: If804b515210a38eb353c1fd65b8f09616607964f
Reviewed-on: https://chromium-review.googlesource.com/1172075
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>