graphics_utils: Optimize quadratic behavior adding failures.
graphics_dEQP adds every test to the failures list for the
first run. For adding, we first checked if a test with the
description already exist by linear scan through the list
of failures. However, with 200k+ tests in vk-master this
results in very bad quadratic behavior.

This adds a dictionary to look up targets by descriptions.
We never modify the description of a target or remove it
(only remove names from the target), so we just have to add
it on target insertion.

BUG=b:111559548
TEST=start graphics_dEQP.vk-master as client test on Grunt

Change-Id: I4f915ad2d4c4a1b143761a52bf0521f965070042
Reviewed-on: https://chromium-review.googlesource.com/1172296
Commit-Ready: Bas Nieuwenhuizen <basni@chromium.org>
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Pohsien Wang <pwang@chromium.org>