New autotest for testing large number of PPD files.
The purpose of this autotest is to verify that given subset of PPD files
work in ChromeOS. Each PPD file is tested with the following procedure:
1. A printer driver is added to CUPS server.
2. Test documents are printed on the configured printer.
3. Raw output from the CUPS server is intercepted by FakePrinter.
4. Obtained outputs are verified (see readme.txt file for details).
5. The printer driver is removed from CUPS server.
This procedure is repeated for every PPD file. The number of PPD files
may be large (3K+ files). To decrease amount of time needed by
the autotest several PPD files are tested simultaneously in parallel
threads.

BUG=chromium:862784
TEST=Tested on ~20 different boards and all 3.2K PPD files

Change-Id: I732bc5e118411258304aa221a0a36f3e8bcc322f
Reviewed-on: https://chromium-review.googlesource.com/1173355
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>