Migrate tests from RETRIES to JOB_RETRIES
RETRIES is an old control file option that triggers retries within
autoserv that are mostly invisible to the rest of the system. This
feature is being killed.
Instead, use JOB_RETRIES that directs the suite that kicks off tests to
retry the test within sane limits.

BUG=chromium:873716
TEST=None

Change-Id: I72c184e8c33cd23367192df7b7bbfdcc16983414
Reviewed-on: https://chromium-review.googlesource.com/1173655
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>