graphics_dEQP: Initialize a test name for bad xml.
If the xml is bad for a test, test_case is not set,
and hence the added failure is with the test_case
name of the previous test.

However, if this happens for the first test,
test_case will not have been defined yet. So this
parses the #beginTestCaseResult stanza to have a
test_case name if the xml is bad. If the xml is
good the name will still be overridden by the xml
provided name.

BUG=b:111559548
TEST=run graphics_dEQP.vk-master on Grunt.

Change-Id: I8b495bb3ecb3122affbb333c21ca99e6944ea900
Reviewed-on: https://chromium-review.googlesource.com/1177386
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Pohsien Wang <pwang@chromium.org>