cr50_test: give cr50 some time to process pwrbtn press
After the last power button press in ccd open, cr50 erases NVMEM and
resets the DUT. After that is complete, it sets the ccd state to open.
Wait 10 seconds in between pressing the power button and checking the
ccd state, so we don't check the state in the middle of this reset
process.

10 seconds is very conservative, but during this poll state press power
button code, cr50 only requesting presses once a minute so waiting 10
seconds is not a big deal.

BUG=none
BRANCH=none
TEST=run firmware_FWMPDisableCCD on nami

Change-Id: I1c3e6b0f4ebb7a3910fb7328fb0fdf67f955b8cf
Reviewed-on: https://chromium-review.googlesource.com/1176702
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>