platform_MemoryPressure: add platform_SimpleMemoryPressure
This adds control.quick, containing the platform_FastMemoryPressure test.
It also tweaks it for faster execution on a 4GB caroline.

The tension here is between making the test faster, but avoiding
allocation so fast that it causes OOM kills.  Further tweaking
may be needed if we start making wider use of this test (for instance,
by monitoring available RAM and slowing down as we get close to 0).

BUG=none
TEST=ran on caroline

Change-Id: I765dc9cf5b26e789275cf6a437aae2ad0c224f0d
Reviewed-on: https://chromium-review.googlesource.com/1180171
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>