tcpdump_analyzer: Allow rejecting low dBm packets.
If the RF shielding in the wificell or other chambers is imperfect,
we'll see packets from the external environment in the packet capture
and tests that assert if the packet capture has certain properties
(i.e. only packets of a certain kind) will fail. A good way to reject
these packets ("leakage from the outside world") is to look at signal
strength. The DUT is usually either next to the AP or <5ft from the AP
in these chambers. A signal strength of < -85 dBm in an incoming packet
should imply it is leakage. This option is turned off by default and
external packets are part of the capture by default. Be careful to not
turn on this option in an attenuated setup, where the DUT/AP packets
will also have a low signal (i.e. network_WiFi_AttenPerf).

BUG=chromium:875450
TEST=Run network_WiFi_DisableRandomMACAddress at my desk with a
Caroline. Fails by default, passes if this filter is turned on.
`test_that -b caroline $DUT_IP network_WiFi_DisableRandomMACAddress \
--autotest_dir=./ --fast`

The diff applied to turn this was in random_mac_address_test_base.py:

         frames = tcpdump_analyzer.get_frames(
                 results[0].local_pcap_path,
-                tcpdump_analyzer.WLAN_PROBE_REQ_ACCEPTOR, reject_bad_fcs=True)
+                tcpdump_analyzer.WLAN_PROBE_REQ_ACCEPTOR, reject_bad_fcs=True,
+                reject_low_signal=True)

Change-Id: I74b82b06a140256757d904d026fc1ec57e69f460
Reviewed-on: https://chromium-review.googlesource.com/1180655
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>