autotest: Create MySQL dump type for replication
This type, intended to be manually run only for the moment, write-locks the
slave DB for the duration (>3 hours) in order to save a guaranteed
replication point. This allows the dump to be used to spin up a new
replica or restore a broken one.
Alternative to CL:1182311

BUG=chromium:871808
TEST=untested

Change-Id: I3740f8d930d136469a02bad48d28b689c61d323c
Reviewed-on: https://chromium-review.googlesource.com/1182319
Reviewed-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>