firmware_ECLidSwitch: Check for keyboard capability.
There's a test case within firmware_ECLidSwitch that checks to see that
no keycodes are generated when the lid is opened and closed.  However,
it was failing if a DUT did not have a keyboard.  This commit simply
updates the test case to check if the DUT has a keyboard before running
the keycode test case.

BUG=b:112974551
BRANCH=None
TEST=Run firmware_ECLidSwitch against nocturne; verify that it passes.

Change-Id: Ia684f45bca66179431be329da8d4712bd4727b5c
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1187569
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>